refactor(contact-us): make the contact us component use the generic step switch component
