+++ components/Main/ContactUs/MessageForm/constants.ts
@@ -1,0 +1,4 @@
+export const BODY = 'body'
+export const EMAIL = 'email'
+export const NAME = 'name'
+export const ID_PREFIX = 'contact-us-message-form-'
